requirements/base.txt
django==4.2.*
psycopg2-binary>=2.9.10
django-environ==0.11.2
django-money>=3,<4
djangorestframework>=3,<4
django-vite>=3,<4
Pillow>=10,<12
requests>=2.31
